requirements.txt
Django
pillow
django-htmx
django-allauth
django-cleanup
django-widget-tweaks
django-crispy-forms
crispy-bootstrap5
psycopg2-binary
requests
geoip2
